Remove all project-specific CMake files and `.gitignore` as part of repository cleanup
Massively overhauled project buildsystem to use cmake to link multiple projects to the same libraries statically and dynamically on user configuration, with debug and release compile options, and the option to compile one project as a individual target or the entire repo at once.

Updated code to use glad 2 loader functions in code, and added helper scripts that can be accessed by git aliases to improve the process of adding new projects to the build target, and many more changes.

Documentation to be added in the next commit.